:construction: wip ir

+++ src/grammar/ast/mod.rs
@@ -5,5 +5,5 @@ pub mod program;
 pub type Identifier = String;
 pub type Type = String;
 pub type Boolean = bool;
-pub type Int = i32;
+pub type Int = u64;
 pub type Str = String;
